requirements.txt
fastapi[all]
uvicorn[standard]
dotenv

langchain
langchain-azure-ai

sqlmodel
pyodbc
pydantic
azure-identity
sqlalchemy